Implement fractal worker thread

+++ src/fractal/Zoomable.java
@@ -5,12 +5,10 @@ public interface Zoomable {
 	/**
 	 * Zooms 2 dimensional object into some
 	 * 	inner rectangular section.
-	 * Coordinates are given as scales where
-	 * 	0 is for start of the object and 1 is the end.
-	 * @param x - Zoom center x coordinate (in range [0, 1])
-	 * @param y - Zoom center y coordinate (in range [0, 1])
-	 * @param scale - How much to scale the object (in range [0, 1])
+	 * @param x - Zoom center x coordinate in pixels.
+	 * @param y - Zoom center y coordinate in pixels.
+	 * @param scale - How much to scale the object (1 will not change, 2 will zoom twice etc.)
 	 */
-	void zoom(double x, double y, double scale);
+	void zoom(int x, int y, double scale);
 
 }
